+++ hslua-marshalling/CHANGELOG.md
@@ -9,3 +9,8 @@ Release pending.
 - Initially created. Contains modules previously found in the
   modules `Foreign.Lua.Peek` and `Foreign.Lua.Push` from
   `hslua-1.3`.
+
+- Removed most functions from the Userdata module, incl. peekAny,
+  pushAny. The functions don't add much value over those in
+  `HsLua.Core.Userdata`. Use UDTypes from hslua-packaging for a
+  more comfortable method of exposing data via userdata values.